+++ src/main/webapp/resources/static/js/admin/admin.js
@@ -1,19 +1,18 @@
 import { callAjax } from "../api/ajax.js";
 
 $(document).ready(function(){
-	reload(1,"/customer/faqWrite");
+	reload("/admin/customer/faq");
 	$("#pageLink > li > .tab_tit").click(handleLinkClick);
 })
 
 function handleLinkClick(){
-	const idx = $(this).data("id");
 	const url = $(this).data("link");
-	reload(idx, url);
+	reload(url);
 	$("#pageLink > li").removeClass("active");
 	$(this).parent().addClass("active");
 }
 
-async function reload(idx=1,url){
+async function reload(url){
 	try{
 		const content = await callAjax('GET',url);
